trac-executors/python/requirements.txt

protobuf >= 3.13.0, < 3.14.0
protoc-wheel-0 >= 3.13.0, < 3.14.0
